Update env fix handling logic
