+++ app/src/main/java/com/topjohnwu/magisk/core/download/NotificationService.kt
@@ -7,16 +7,14 @@ import com.topjohnwu.magisk.core.base.BaseService
 import com.topjohnwu.magisk.core.view.Notifications
 import org.koin.core.KoinComponent
 import java.util.*
+import kotlin.collections.HashMap
 import kotlin.random.Random.Default.nextInt
 
 abstract class NotificationService : BaseService(), KoinComponent {
 
-    abstract val defaultNotification: Notification.Builder
-
     private val hasNotifications get() = notifications.isNotEmpty()
 
-    private val notifications =
-        Collections.synchronizedMap(mutableMapOf<Int, Notification.Builder>())
+    private val notifications = Collections.synchronizedMap(HashMap<Int, Notification.Builder>())
 
     override fun onTaskRemoved(rootIntent: Intent?) {
         super.onTaskRemoved(rootIntent)
@@ -24,22 +22,23 @@ abstract class NotificationService : BaseService(), KoinComponent {
         notifications.clear()
     }
 
+    abstract fun createNotification(): Notification.Builder
+
     // --
 
     fun update(
         id: Int,
         body: (Notification.Builder) -> Unit = {}
     ) {
-        val notification = notifications.getOrPut(id) { defaultNotification }
-
-        notify(id, notification.also(body).build())
-
-        if (notifications.size == 1) {
+        val wasEmpty = notifications.isEmpty()
+        val notification = notifications.getOrPut(id, ::createNotification).also(body)
+        if (wasEmpty)
             updateForeground()
-        }
+        else
+            notify(id, notification.build())
     }
 
-    protected fun finishNotify(
+    protected fun lastNotify(
         id: Int,
         editBody: (Notification.Builder) -> Notification.Builder? = { null }
     ) : Int {
@@ -57,6 +56,11 @@ abstract class NotificationService : BaseService(), KoinComponent {
         return newId
     }
 
+    protected fun remove(id: Int) = notifications.remove(id).also {
+        cancel(id)
+        updateForeground()
+    }
+
     // ---
 
     private fun notify(id: Int, notification: Notification) {
@@ -67,16 +71,13 @@ abstract class NotificationService : BaseService(), KoinComponent {
         Notifications.mgr.cancel(id)
     }
 
-    protected fun remove(id: Int) = notifications.remove(id).also {
-        cancel(id)
-        updateForeground()
-    }
-
     private fun updateForeground() {
-        if (hasNotifications)
-            startForeground(notifications.keys.first(), notifications.values.first().build())
-        else
+        if (hasNotifications) {
+            val first = notifications.entries.first()
+            startForeground(first.key, first.value.build())
+        } else {
             stopForeground(true)
+        }
     }
 
     // --
