+++ app/src/main/java/com/topjohnwu/magisk/core/tasks/FlashResultListener.kt
@@ -1,0 +1,10 @@
+package com.topjohnwu.magisk.core.tasks
+
+import androidx.annotation.MainThread
+
+interface FlashResultListener {
+
+    @MainThread
+    fun onResult(success: Boolean)
+
+}
