+++ app/src/main/java/com/topjohnwu/magisk/model/entity/internal/Configuration.kt
@@ -32,6 +32,9 @@ sealed class Configuration : Parcelable {
     object Uninstall : Configuration()
 
     @Parcelize
+    object EnvFix : Configuration()
+
+    @Parcelize
     data class Patch(val fileUri: Uri) : Configuration()
 
 }
