--- app/src/main/java/com/topjohnwu/magisk/model/flash/FlashResultListener.kt
@@ -1,7 +1,0 @@
-package com.topjohnwu.magisk.model.flash
-
-interface FlashResultListener {
-
-    fun onResult(isSuccess: Boolean)
-
-}
